Set key to empty when S3OpenWrite closed writing any data
S3 does not support zero-content multi-part file uploads, i.e. if you try to
multi-part upload no data then you'll get an error. This meant that if, for
some reason, you opened an S3 key for writing but actually didn't write to it
at all then nothing would be written to that key.

This changes S3OpenWrite to fallback to set_contents_from_string to set a key
to have an empty (zero byte) in this situation.