Remove explicit version pins
It's generally good practice to be as tolerant as possible in libs like this.
Unless there is a very good reason not to, don't put pins on dependencies. I'm
assuming this uses specific httpretty>=0.8.10 features and boto>=2.32 features.
Also, it's good practice to only ever use `>=`, and never `<=` (as this will
prevent these dependencies from being upgraded).