Refactoring smart_open to share compression and encoding functionality (#185)
* Refactoring smart_open to share compression and encoding functionality

All of our transport methods (file, S3, HDFS, WebHDFS) now read and
write bytes.  A shared compression layer sits over that and performs
compression and decompression transparently.  A shared encoding layer
sits over that, and performs compression decompression transparently.

The benefit of this approach is that it decouples actual I/O,
compression and encoding into independent layers.  The I/O layers now
have to worry only about binary I/O.  The compression all happens in one
place, so adding new codecs is simple.  Finally, encoding also happens
in one place, with the same benefits.

Other things I did:

- ripped out S3 text I/O, we do not need this anymore
- rewrote HDFS as a IOBase-based separate module
- split http module
- rewrote WebHDFS subsystem based on io.IOBase
- get rid of some unused imports

* fixup in webhdfs.py: detect Py2 correctly

* Disable gzip tests for Py2

This needs a separate implementation of seek to work under Py2.
gzip with hdfs/http wasn't supported at all prior to the refactoring, so
it is a separate issue.

* fix unit tests by explicitly specifying encoding

By default, the system encoding is used when opening a file.
The tests expect the encoding to be UTF-8, so if the system encoding
happens to be anything else, the tests will fail.

Some Py2 installations use ascii as the default encoding.

* adding HTTP integration tests

* when seek is missing but required, buffer in memory

* work around missing .seekable in py2

* fixup in webhdfs.py: expect response.content to be bytes

* added some sample code for WebHDFS/HDFS integration tests

* specify working directory for HTTP server

* include http tests in travis.yml

* sleep for 1s to avoid race condition

* fixup in http integration tests

* set Accept-Encoding header, point tests at github.com