Adding dummy method for flush.
We need this method in order for S3 to mimic local files.  According to
the docs, this method must:

> Flush the write buffers of the stream if applicable. This does nothing
> for read-only and non-blocking streams.

Since we're doing a multipart upload to S3, it's impossible to flush the
buffers.  The flushed content would not be available until the user
calls the close method.

So, in our case, the flush method does nothing.