Allow zero as size for private S3 buffer methods
Change the tests in the size arg default-setting conditions for
smart_open.s3.BufferedInputBase from `size > 0` to `size >= 0`. The
first one could be a genuine bug (though it was never called this way by
the class), because when a caller would ask for zero bytes, they'd get
up to a whole buffer_size's worth back. The second one is a semantics
change: if a caller really wants to fill the buffer with zero bytes,
we'll let them.