Support filling ByteBuffers from lists
It would be beneficial to fill ByteBuffers with lists that contain
bytestrings, in order to support experimentation at the REPL and to make
doctest examples more succinct. To that end, change the implementation
of its `fill` method to use a for loop to iterate over the byte source
when it's not a file-like object, and break out of the loop once the
condition is met (rather than a while loop that uses `next`, which
doesn't work on lists).

Now that `fill` accepts lists in addition to file-like objects and
iterators, change the byte source argument name from the clunky
`reader_or_iterable` to `source`, and explain what it's allowed to be in
the parameter section of the docstring.

Add a test to ensure that filling from lists works.