Add usage example/doctest to ByteBuffer docstring
This only passes as a doctest in Python 3 because of the bytestring
literals for some of the expected values. In Python 2, those values are
simply regular strings and doctest seems to work by comparing the string
representations of the values rather than the values themselves, so
those lines fail because they don't have the `b` prefix.