Benchmark many small reads on a large buffer
Add the `test_s3_performance_small_reads` benchmark to the S3
integration tests. This benchmark creates a test file one megabyte in
size that's filled with 65,536 (64 KiB) "messages", each prefixed with
a byte indicating their length. The file is downloaded into a buffer of
the same size, and then each length byte and message is read one at
a time from the buffer, for a total of 131,072 `read()` calls.

Before the changes in this PR, downloading and reading the file this way
took 4.97 seconds on average. After this PR's changes, it takes only
1.61 seconds, a 68% improvement in performance. The magnitude of
improvement increases as the size of the buffer increases, or the size
of the messages decreases.