Implement option to write to a custom buffer during S3 writes (#547)
* Implement option to buffer to disk for S3 writes

Currently, we buffer in-memory. This works well in most cases, but can
be a problem when performing many uploads in parallel.  Each upload
needs to buffer at least 5MB of data (that is the minimum part size for
AWS S3). When performing hundreds or thousands of uploads, memory can
become a bottleneck.

This PR relieves that bottleneck by buffering to temporary files
instead. Each part gets saved to a separate temporary file and discarded
after upload. This reduces memory usage at the cost of increased IO with
the local disk.

* update help.txt

* improve logging message

* simplify implementation

re-use the same buffer in the same writer instance

* dedent

* update help.txt

* fixup

* update howto.md

* Update CHANGELOG.md