Fix error when using GCP urls without deps (#593)
* Fix error when using GCP urls without deps

Currently if you run code like this:

    from smart_open import open
    open("gs://asdf")

This is a fake URL but it works for this problem as is. Without the
Google Cloud deps, you get an error that says to do this:

    pip install smart_open[gcs]

But that doesn't work. You have to do this:

    pip install smart_open[gcp]

It seemed easier to change setup.py so that's what I've done here.

* Update setup.py

Co-authored-by: Michael Penkov <m@penkov.dev>

* Remove remaining GCP references

Co-authored-by: Michael Penkov <m@penkov.dev>